Silences 32bit MSVC warning in tcpip code.
Port number was represented by long rather than unsigned short which has the correct range.

+++ tiberiandawn/tcpip.cpp
@@ -60,7 +60,7 @@
 
 /* these are temporary to allow linking, this manager class is not used at the moment -hifi */
 char PlanetWestwoodIPAddress[IP_ADDRESS_MAX];
-long PlanetWestwoodPortNumber;
+unsigned short PlanetWestwoodPortNumber;
 bool PlanetWestwoodIsHost;
 
 /*
